# Refactor Client Cache