--- a/Uchu.Core/Client/Attribute/CacheIndexAttribute.cs
+++ b/Uchu.Core/Client/Attribute/CacheIndexAttribute.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Uchu.Core.Client.Attribute
+{
+    [AttributeUsage(AttributeTargets.Property)]
+    public class CacheIndexAttribute : System.Attribute
+    {
+        
+    }
+}