--- a/Uchu.Core/Client/Models/BehaviorTemplate.cs
+++ b/Uchu.Core/Client/Models/BehaviorTemplate.cs
@@ -1,15 +1,17 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
+using Uchu.Core.Client.Attribute;
 
 namespace Uchu.Core.Client
 {
 	[Table("BehaviorTemplate")]
+	[CacheMethod(CacheMethod.Persistent)]
 	public class BehaviorTemplate
 	{
 		[Key] [Column("efId")]
 		public int EfId { get; set; }
 
-		[Column("behaviorID")]
+		[CacheIndex] [Column("behaviorID")]
 		public int? BehaviorID { get; set; }
 
 		[Column("templateID")]