--- a/Uchu.Core/Client/Models/DestructibleComponent.cs
+++ b/Uchu.Core/Client/Models/DestructibleComponent.cs
@@ -1,15 +1,17 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
+using Uchu.Core.Client.Attribute;
 
 namespace Uchu.Core.Client
 {
 	[Table("DestructibleComponent")]
+	[CacheMethod(CacheMethod.Burst)]
 	public class DestructibleComponent
 	{
 		[Key] [Column("efId")]
 		public int EfId { get; set; }
 
-		[Column("id")]
+		[CacheIndex] [Column("id")]
 		public int? Id { get; set; }
 
 		[Column("faction")]