--- a/Uchu.Core/Client/Models/ItemComponent.cs
+++ b/Uchu.Core/Client/Models/ItemComponent.cs
@@ -1,15 +1,17 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
+using Uchu.Core.Client.Attribute;
 
 namespace Uchu.Core.Client
 {
 	[Table("ItemComponent")]
+	[CacheMethod(CacheMethod.Burst)]
 	public class ItemComponent
 	{
 		[Key] [Column("efId")]
 		public int EfId { get; set; }
 
-		[Column("id")]
+		[CacheIndex] [Column("id")]
 		public int? Id { get; set; }
 
 		[Column("equipLocation")]