--- a/Uchu.Core/Client/Models/Objects.cs
+++ b/Uchu.Core/Client/Models/Objects.cs
@@ -1,15 +1,17 @@
 using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
+using Uchu.Core.Client.Attribute;
 
 namespace Uchu.Core.Client
 {
 	[Table("Objects")]
+	[CacheMethod(CacheMethod.Burst)]
 	public class Objects
 	{
 		[Key] [Column("efId")]
 		public int EfId { get; set; }
 
-		[Column("id")]
+		[CacheIndex] [Column("id")]
 		public int? Id { get; set; }
 
 		[Column("name")]