--- a/Uchu.StandardScripts/ForbiddenValley/FriendOfTheNinja.cs
+++ b/Uchu.StandardScripts/ForbiddenValley/FriendOfTheNinja.cs
@@ -18,8 +18,7 @@ namespace Uchu.StandardScripts.ForbiddenValley
             {
                 // Load physics object information from cdclient
                 var phantomPhysicsComponentId = gameObject.Lot.GetComponentId(ComponentId.PhantomPhysicsComponent);
-                var cdcComponent = ClientCache.GetTable<Core.Client.PhysicsComponent>()
-                    .FirstOrDefault(r => r.Id == phantomPhysicsComponentId);
+                var cdcComponent = ClientCache.Find<Core.Client.PhysicsComponent>(phantomPhysicsComponentId);
                 var assetPath = cdcComponent?.Physicsasset;
 
                 // Configure physics object