… Uchu.World/Client/CdClient/CacheTable/BaseTableCache.cs
… Uchu.World/Client/CdClient/CacheTable/BurstCacheTable.cs
… Uchu.World/Client/CdClient/CacheTable/EntryCacheTable.cs
… Uchu.World/Client/CdClient/CacheTable/PersistentCacheTable.cs
… Uchu.World/Client/CdClient/ClientCache.cs
… Uchu.World/Client/CdClient/EntryCacheTable.cs
… Uchu.World/Client/CdClient/PersistentCacheTable.cs
… Uchu.World/Client/CdClient/TableCache.cs
… Uchu.World/Client/ZoneParser.cs
… Uchu.World/Handlers/Commands/CharacterCommandHandler.cs
… Uchu.World/Handlers/GameMessages/GeneralHandler.cs
… Uchu.World/Handlers/WorldInitializationHandler.cs
… Uchu.World/Objects/Components/DestroyableComponent.cs
… Uchu.World/Objects/Components/Player/InventoryManagerComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/BaseCombatAIComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/CharacterComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/DestructibleComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/InventoryComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/LuaScriptComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/MovementAiComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/PetComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/PhantomPhysicsComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/QuickBuildComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/RailActivatorComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/ScriptedActivityComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/SkillComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/TriggerComponent.cs
… Uchu.World/Objects/Components/ReplicaComponents/VendorComponent.cs
… Uchu.World/Objects/Components/Server/LootContainerComponent.cs
… Uchu.World/Objects/Components/Server/MissionGiverComponent.cs