This pull request attempts to improve how caching of the client database is handled. Rather than running direct queries, most of which are just checking an id, a query is done to the cache for a specific id. There are 3 types of caching methods:
- Entry: The database is queried for every request for an id that isn't cached.
- Burst: The entire database is loaded into memory for up to a few seconds after the last call to it.
- Persistent: The entire database is loaded into memory permanently. This is currently only done for behaviors because of how long it takes to run queries, which results in the first of a behavior type being either slow or not work at all.

When an entire table is loaded into memory, it is first converted to a dictionary by the specified index. This is done to make finding entries with a given id a dictionary read instead of iterating over a list. From my testing:
- Starting Avant Gardens with my larger character (completed missions up to Nexus Tower with full Rank 2 armor) went from **14.867 seconds to 9.513 seconds**, measured from the first message logged to "<Player> loaded...".
- Starting Avant Gardens with my smaller character (completed missions up to Block Yard) went from **13.487 seconds to 9.534 seconds**, measured from the first message logged to "<Player> loaded...".
- Performing skills for the first time, such as equipping items, smashing Venture Explorer crates, and scripts like the monument lasers, are *much* faster.
- Performing queries is simpler to write now since it is just "find the type that matches the id" instead of "perform this LINQ statement that returns if the ids match".

Memory usage is a bit mixed from my testing. The average memory usage is significantly higher on world load, but the peak memory usage is significantly lower. Idle memory usage was about 240MB vs 310MB before and after the cache. However, I noticed peak memory usage going all the way up to 370MB without the caching changes, and never going above about 320MB with the cache changes.